% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{apply_column_spliceview}
\alias{apply_column_spliceview}
\title{Apply summary function to the columns in each segment of a SpliceView object}
\usage{
apply_column_spliceview(sv, FUN, simplify = FALSE, USE.NAMES = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Apply summary function to each data point column in a SplicedView and return list of output data.
}
