% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{apply_filter}
\alias{apply_filter}
\title{Apply a filter to a View}
\usage{
apply_filter(
  view,
  data_points,
  sig_filter,
  param_str = "",
  folder_out = "Filtered",
  save_output = FALSE
)
}
\arguments{
\item{view}{\code{ProcessedView} object.}

\item{data_points}{body parts e.g. 'Nose'.}

\item{sig_filter}{S3 filter object from signals package.}

\item{param_str}{string of parameter values to add to output file if desired.}

\item{folder_out}{output folder relative to recording home (default is 'Filtered').}

\item{save_output}{save the output?}
}
\value{
a filtered object.
}
\description{
Apply a filter to a View
}
