% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{apply_segment_spliceview}
\alias{apply_segment_spliceview}
\title{Apply complex function to each segment in a SpliceView object}
\usage{
apply_segment_spliceview(sv, FUN, ...)
}
\arguments{
\item{sv}{\code{SplicedView} object.}

\item{FUN}{function to apply.}

\item{...}{passed to FUN.}
}
\value{
list of two elements: 'output' containing results of apply FUN to 'input'
}
\description{
Apply complex function to each segment in a SpliceView object
}
