% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{ave_power_over_splices}
\alias{ave_power_over_splices}
\title{Calculate mean average power over splices using a splicing table}
\usage{
ave_power_over_splices(
  jv,
  splicing_df,
  num_splices,
  column,
  sampling_type = "offset",
  rejection_list = list(),
  include_original = TRUE,
  show_plot = TRUE
)
}
\arguments{
\item{num_splices}{}
}
\description{
Randomly generates splices from a splicing table and calculates average
power for each segment and splice. Calculates the mean average power
over the random splices for each segment and period. Compares with the
average power for the original splice.
}
