% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{compare_ave_power1}
\alias{compare_ave_power1}
\title{Compare average power distribution using a splicing table}
\usage{
compare_ave_power1(
  jv,
  splicing_df,
  splice_name,
  num_segment_samples,
  num_splice_samples,
  column,
  sampling_type = "offset",
  rejection_list = list(),
  show_plot = TRUE
)
}
\arguments{
\item{sampling_type}{}
}
\description{
Compare average power distribution using a splicing table
}
