% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{compare_avg_cross_power2}
\alias{compare_avg_cross_power2}
\title{Compare the average cross power distribution of two SplicedViews using
sampling on each segment}
\usage{
compare_avg_cross_power2(
  sv1,
  sv2,
  name1,
  name2,
  num_samples,
  columns,
  show_plot = TRUE
)
}
\arguments{
\item{show_plot}{}
}
\description{
Compare the average cross power distribution of two SplicedViews using
sampling on each segment
}
