% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asynchrony.R
\name{difference_onsets}
\alias{difference_onsets}
\title{Get onset differences}
\usage{
difference_onsets(onset_obj, instruments, expr = NULL, splicing_dfr = NULL)
}
\arguments{
\item{onset_obj}{\code{OnsetsSelected} object.}

\item{instruments}{character vector of instrument names.}

\item{expr}{R expression to subset onsets (Default is NULL).}

\item{splicing_dfr}{\code{Splice} object (Default is NULL).}
}
\value{
\code{OnsetsDifference} object.
}
\description{
Calculates the difference in onset times for each instrument pair in
milli-seconds.
}
\examples{
r1 <- get_sample_recording()
o1 <- get_onsets_selected_data(r1)
head(difference_onsets(o1, instruments = c('Inst', 'Tabla')))
head(difference_onsets(o1, instruments = c('Inst', 'Tabla'), expr = Matra == 3))
}
\seealso{
Other asynchrony analysis functions: 
\code{\link{summary_onsets}()}
}
\concept{asynchrony analysis functions}
