% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data_points}
\alias{get_data_points}
\title{Get the data points held in a view}
\usage{
get_data_points(obj)
}
\arguments{
\item{obj}{\code{View} object.}
}
\value{
character vector of body parts.
}
\description{
Get the data points held in a view
}
\examples{
r <- get_sample_recording()
rv <- get_raw_view(r, "Central", "", "Sitar")
get_data_points(rv)
}
\seealso{
Other data functions: 
\code{\link{apply_filter_sgolay}()},
\code{\link{get_duration_annotation_data}()},
\code{\link{get_feature_data}()},
\code{\link{get_filtered_views}()},
\code{\link{get_joined_view}()},
\code{\link{get_metre_data}()},
\code{\link{get_onsets_selected_data}()},
\code{\link{get_processed_views}()},
\code{\link{get_processed_view}()},
\code{\link{get_raw_optflow_view}()},
\code{\link{get_raw_views}()},
\code{\link{get_raw_view}()},
\code{\link{get_recording}()},
\code{\link{get_sample_recording}()}
}
\concept{data functions}
