% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_onsets_selected_data}
\alias{get_onsets_selected_data}
\title{Get onsets selected files}
\usage{
get_onsets_selected_data(recording)
}
\arguments{
\item{recording}{\code{Recording} object.}
}
\value{
list of data.frames
}
\description{
Get onsets selected files
}
\examples{
r <- get_recording("NIR_ABh_Puriya", fps = 25)
o <- get_onsets_selected_data(r)
}
\seealso{
Other data functions: 
\code{\link{apply_filter_sgolay}()},
\code{\link{get_data_points}()},
\code{\link{get_duration_annotation_data}()},
\code{\link{get_feature_data}()},
\code{\link{get_filtered_views}()},
\code{\link{get_joined_view}()},
\code{\link{get_metre_data}()},
\code{\link{get_processed_views}()},
\code{\link{get_processed_view}()},
\code{\link{get_raw_optflow_view}()},
\code{\link{get_raw_views}()},
\code{\link{get_raw_view}()},
\code{\link{get_recording}()},
\code{\link{get_sample_recording}()}
}
\concept{data functions}
