% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{get_osf_recording}
\alias{get_osf_recording}
\title{Get movementsync recording from OSF}
\usage{
get_osf_recording(
  stem = "Original",
  to_dir = "~/movementsync/Original",
  overwrite = FALSE
)
}
\arguments{
\item{stem}{zip file stem (default is 'Original' which contains all the
recordings for the walk-throughs.}

\item{to_dir}{directory to copy to (default is "~/movementsync/Original").}

\item{overwrite}{overwriting existing dataset files?}
}
\value{
invisible vector of CSV file names.
}
\description{
Get movementsync recording from OSF
}
\examples{
\dontrun{
get_osf_recording()
}
}
