% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{get_osf_recordings}
\alias{get_osf_recordings}
\title{Get movementsync recording from OSF}
\usage{
get_osf_recordings(
  stems,
  node,
  to_dir = "~/movementsync/Original",
  overwrite = FALSE
)
}
\arguments{
\item{stems}{vector of zip file stem(s).}

\item{node}{OSF node for movementsync.}

\item{to_dir}{directory to copy to (default is "~/movementsync/Original").}

\item{overwrite}{overwriting existing dataset files?}
}
\value{
invisible vector of downloaded CSV file names.
}
\description{
Get movementsync recording from OSF
}
\examples{
\dontrun{
get_osf_recordings()
}
}
