% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asynchrony.R
\name{get_pair_columns}
\alias{get_pair_columns}
\title{Helper for ggpairs plots}
\usage{
get_pair_columns(
  obj,
  instrument_pairs,
  instrument_names,
  add_ref_beat_time = FALSE,
  add_segment = FALSE,
  add_metre = FALSE,
  partial = FALSE
)
}
\arguments{
\item{obj}{'OnsetsDifference' object.}

\item{instrument_pairs}{character vector of instrument pairs or
a single instrument pair for partial matching.}

\item{instrument_names}{character vector of instrument names}

\item{add_ref_beat_time}{add a Reference_Beat_Time column? (Default is FALSE).}

\item{add_segment}{add a Segment column? (Default is FALSE).}

\item{add_metre}{add a Metre column? (Default is FALSE).}

\item{partial}{should matching be partial? (Default is FALSE)}
}
\value{
numeric vector of column positions in the \code{OnsetsDifference} object
}
\description{
Finds the column numbers of a subset of instrument pairs in a
\code{OnsetsDifference} object using partial matching.
}
\examples{
\dontrun{
library(GGally)
instruments <- c("Shoko_L", "Shoko_R", "Taiko", "Kakko", "Kakko_1", "So", "Biwa",
"Ryuteki", "Hichiriki", "Sho", "Biwa_RW", "Shoko_RW", "Taiko_LW", "Taiko_RW")
r5 <- get_recording("Gagaku_5_Juha", fps = 60)
o5 <- get_onsets_selected_data(r5)
po5 <- difference_onsets(o5, instruments = instruments)
pair_cols <- get_pair_columns(po5, c("Shoko_L-Kakko", "So-Biwa"), instruments)
ggpairs(po5, columns = pair_cols)
}
}
\seealso{
Other asynchrony analysis functions: 
\code{\link{difference_onsets}()},
\code{\link{summary_onsets}()}
}
\concept{asynchrony analysis functions}
