% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_processed_views}
\alias{get_processed_views}
\title{Get processed views}
\usage{
get_processed_views(r, data_points, filetype = "rda")
}
\arguments{
\item{r}{\code{Recording} object.}

\item{data_points}{vector of body parts e.g. 'Nose'.}

\item{filetype}{type of file ('rda' as default), can be 'csv'.}
}
\value{
list of \code{ProcessedView} objects.
}
\description{
Get processed views
}
\examples{
r <- get_sample_recording()
pv_list <- get_processed_views(r)
plot(pv_list$Central_Tabla)
}
\seealso{
Other data functions: 
\code{\link{apply_filter_sgolay}()},
\code{\link{get_data_points}()},
\code{\link{get_duration_annotation_data}()},
\code{\link{get_feature_data}()},
\code{\link{get_filtered_views}()},
\code{\link{get_joined_view}()},
\code{\link{get_metre_data}()},
\code{\link{get_onsets_selected_data}()},
\code{\link{get_processed_view}()},
\code{\link{get_raw_optflow_view}()},
\code{\link{get_raw_views}()},
\code{\link{get_raw_view}()},
\code{\link{get_recording}()},
\code{\link{get_sample_recording}()}
}
\concept{data functions}
