% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_raw_optflow_view}
\alias{get_raw_optflow_view}
\title{Creates time reference and displacement from raw csv optflow data}
\usage{
get_raw_optflow_view(
  recording,
  vid,
  direct,
  inst,
  folder_out = "Raw",
  save_output = TRUE
)
}
\arguments{
\item{recording}{\code{Recording} object.}

\item{vid}{camera.}

\item{direct}{direction.}

\item{inst}{instrument.}

\item{folder_out}{output folder relative to recording home  (default is 'Raw').}

\item{save_output}{save the output?}
}
\value{
an \code{OptFlowView} object.
}
\description{
Used to load OptFlow data.
}
\examples{
\dontrun{
r <- get_recording("NIR_DBh_Malhar_2Gats", fps = 25)
rov <- get_raw_optflow_view(r, "" ,"", "Guitar")
pov <- get_processed_view(rov)
fv1 <- apply_filter_sgolay(pov, c("Head"), n=19, p=4)
}
}
\seealso{
Other data functions: 
\code{\link{apply_filter_sgolay}()},
\code{\link{get_data_points}()},
\code{\link{get_duration_annotation_data}()},
\code{\link{get_feature_data}()},
\code{\link{get_filtered_views}()},
\code{\link{get_joined_view}()},
\code{\link{get_metre_data}()},
\code{\link{get_onsets_selected_data}()},
\code{\link{get_processed_views}()},
\code{\link{get_processed_view}()},
\code{\link{get_raw_views}()},
\code{\link{get_raw_view}()},
\code{\link{get_recording}()},
\code{\link{get_sample_recording}()}
}
\concept{data functions}
