% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_raw_views}
\alias{get_raw_views}
\title{Get Pose views from a recording}
\usage{
get_raw_views(recording, filetype = "rda")
}
\arguments{
\item{recording}{\code{Recording} object.}

\item{filetype}{type of file ('rda' as default), can be 'csv'.}
}
\value{
named list of views
}
\description{
Get Pose views from a recording
}
\examples{
r <- get_sample_recording()
v_list <- get_raw_views(r)
}
\seealso{
Other data functions: 
\code{\link{apply_filter_sgolay}()},
\code{\link{get_data_points}()},
\code{\link{get_duration_annotation_data}()},
\code{\link{get_feature_data}()},
\code{\link{get_filtered_views}()},
\code{\link{get_joined_view}()},
\code{\link{get_metre_data}()},
\code{\link{get_onsets_selected_data}()},
\code{\link{get_processed_views}()},
\code{\link{get_processed_view}()},
\code{\link{get_raw_optflow_view}()},
\code{\link{get_raw_view}()},
\code{\link{get_recording}()},
\code{\link{get_sample_recording}()}
}
\concept{data functions}
