% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{is_splice_overlapping}
\alias{is_splice_overlapping}
\title{Checks if splicing data.frames overlap}
\usage{
is_splice_overlapping(...)
}
\arguments{
\item{...}{Each argument can be a data frame or a list of data frames}
}
\value{
logical
}
\description{
Checks if splicing data.frames overlap
}
\examples{
l1 <- list(a=c(1, 10), a = c(20, 30), b = c(30, 40))
dfr1 <- splice_time(l1)
l2 <- list(a=c(10, 15), b = c(15, 25))
dfr2 <- splice_time(l2)
is_splice_overlapping(dfr1, dfr2)
}
\seealso{
Other splicing functions: 
\code{\link{clip_splice}()},
\code{\link{get_spliced_view}()},
\code{\link{merge_splice}()},
\code{\link{splice_time.Duration}()},
\code{\link{splice_time.Metre}()},
\code{\link{splice_time.OnsetsDifference}()},
\code{\link{splice_time.View}()},
\code{\link{splice_time.list}()},
\code{\link{splice_time}()},
\code{\link{split.SplicedView}()}
}
\concept{splicing functions}
