% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{list_local_recordings}
\alias{list_local_recordings}
\title{List downloaded recordings}
\usage{
list_local_recordings(folder_in = "Original", path = "~/movementsync")
}
\arguments{
\item{folder_in}{input folder relative to recording home (default is 'Original').}

\item{path}{recording home folder.}
}
\value{
character vector of stem names
}
\description{
List downloaded recordings
}
\examples{
list_local_recordings()
}
