% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{list_osf_recordings}
\alias{list_osf_recordings}
\title{List available recordings for movementsync from OSF}
\usage{
list_osf_recordings(node = "https://osf.io/w2s3a")
}
\arguments{
\item{node}{OSF node for movementsync.}
}
\value{
character vector of stem names
}
\description{
List available recordings for movementsync from OSF
}
\examples{
\dontrun{
list_osf_recordings()
}
}
