% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIR_ABh_Puriya_Annotation.R
\docType{data}
\name{NIR_ABh_Puriya_Annotation}
\alias{NIR_ABh_Puriya_Annotation}
\title{NIR_ABh_Puriya_Annotation}
\format{
\subsection{\code{rda}}{

A data frame with 161 rows and 5 columns:
\describe{
\item{START-END}{Type of annotation}
\item{2nd colum}{Onset of annotation}
\item{3rd colum}{Offset of annotation}
\item{4th colum}{Duration of annotation}
\item{5th colum}{Description}
...
}
}
}
\source{
\url{https://osf.io/tj2n5}
}
\usage{
data(NIR_ABh_Puriya_Annotation)
}
\description{
A subset of data from NIR_ABh_Puriya annotation.
The data comes from a collection of audiovisual recordings of
North Indian (Hindustani) raga performances which are
part of IEMP North Indian Raga collection, collected and
curated by Martin Clayton, Laura Leante, and Simone Tarsitani.
}
\keyword{datasets}
