% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIR_ABh_Puriya_Annotation_Influence.R
\docType{data}
\name{NIR_ABh_Puriya_Annotation_Influence}
\alias{NIR_ABh_Puriya_Annotation_Influence}
\title{NIR_ABh_Puriya_Annotation_Influence}
\format{
\subsection{\code{rda}}{

A data frame with 306 rows and 5 columns:
\describe{
\item{Event}{Type of event}
\item{Onset time}{Start of the event in seconds}
\item{Offset time}{End of the event in seconds}
\item{Duration}{Duration of the event in seconds}
\item{Notes}{Text notes}
...
}
}
}
\source{
\url{https://osf.io/ks325/}
}
\usage{
data(NIR_ABh_Puriya_Annotation_Influence)
}
\description{
A subset of data from NIR_ABh_Puriya describing
the annotated influence.
The data comes from a collection of audiovisual recordings of
North Indian (Hindustani) raga performances which are
part of IEMP North Indian Raga collection, collected and
curated by Martin Clayton, Laura Leante, and Simone Tarsitani.
}
\keyword{datasets}
