% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIR_ABh_Puriya_Central_Feature_Sitar.R
\docType{data}
\name{NIR_ABh_Puriya_Central_Feature_Sitar}
\alias{NIR_ABh_Puriya_Central_Feature_Sitar}
\title{NIR_ABh_Puriya_Central_Feature_Sitar}
\format{
\subsection{\code{rda}}{

A data frame with 1,501 rows and 3 columns:
\describe{
\item{X}{Frame (here 25 fps)}
\item{Pitch}{Pitch in Hz - Dummy data}
\item{Smooth}{Smooth - Dummy data}
...
}
}
}
\source{
\url{https://osf.io/tj2n5}
}
\usage{
data(NIR_ABh_Puriya_Central_Feature_Sitar)
}
\description{
A subset of data from NIR_ABh_Puriya describing
sitar pitch. Dummy data for demonstration purposes.
The data comes from a collection of audiovisual recordings of
North Indian (Hindustani) raga performances which are
part of IEMP North Indian Raga collection, collected and
curated by Martin Clayton, Laura Leante, and Simone Tarsitani.
}
\keyword{datasets}
