% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIR_ABh_Puriya_Metre_DrutTeental.R
\docType{data}
\name{NIR_ABh_Puriya_Metre_DrutTeental}
\alias{NIR_ABh_Puriya_Metre_DrutTeental}
\title{NIR_ABh_Puriya_Metre_DrutTeental}
\format{
\subsection{\code{rda}}{

A data frame with 351 rows and 3 columns:
\describe{
\item{Cycle}{Number of the Cycle}
\item{Time}{Time in seconds}
\item{Notes}{text which is empty for this file}
...
}
}
}
\source{
\url{https://osf.io/fzv3k}
}
\usage{
data(NIR_ABh_Puriya_Metre_DrutTeental)
}
\description{
A subset of data from NIR_ABh_Puriya Describing
Metre (Cycle numbers and onset times).
The data comes from a collection of audiovisual recordings of
North Indian (Hindustani) raga performances which are
part of IEMP North Indian Raga collection, collected and
curated by Martin Clayton, Laura Leante, and Simone Tarsitani.
}
\keyword{datasets}
