% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIR_ABh_Puriya_Metre_VilambitTeental.R
\docType{data}
\name{NIR_ABh_Puriya_Metre_VilambitTeental}
\alias{NIR_ABh_Puriya_Metre_VilambitTeental}
\title{NIR_ABh_Puriya_Metre_VilambitTeental}
\format{
\subsection{\code{rda}}{

A data frame with 72 rows and 4 columns:
\describe{
\item{Cycle}{Number of the Cycle}
\item{Time}{Time in seconds}
\item{Notes}{text which is empty for this file}
\item{Beats}{Number of beats in the cycle}
...
}
}
}
\source{
\url{https://osf.io/dyu68}
}
\usage{
data(NIR_ABh_Puriya_Metre_VilambitTeental)
}
\description{
A subset of data from NIR_ABh_Puriya describing
the metre in Vilambit Teental section.
}
\details{
The data comes from a collection of audiovisual recordings of
North Indian (Hindustani) raga performances which are
part of IEMP North Indian Raga collection, collected and
curated by Martin Clayton, Laura Leante, and Simone Tarsitani.
}
\keyword{datasets}
