% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIR_ABh_Puriya_OptFlow_Central_Sitar.R
\docType{data}
\name{NIR_ABh_Puriya_OptFlow_Central_Sitar}
\alias{NIR_ABh_Puriya_OptFlow_Central_Sitar}
\title{NIR_ABh_Puriya_OptFlow_Central_Sitar}
\format{
\subsection{\code{rda}}{

A data frame with 1,501 rows and 4 columns:
\describe{
\item{Frame}{Frame (integer, related 25 fps)}
\item{Time}{Time in seconds}
\item{X}{X coordinate}
\item{Y}{Y coordinate}
...
}
}
}
\source{
\url{https://osf.io/r4xza}
}
\usage{
data(NIR_ABh_Puriya_OptFlow_Central_Sitar)
}
\description{
A subset of data from NIR_ABh_Puriya describing
the head movement of the sitar player extracted
using Optical Flow giving X and Y coordinates.
The data comes from a collection of audiovisual recordings of
North Indian (Hindustani) raga performances which are
part of IEMP North Indian Raga collection, collected and
curated by Martin Clayton, Laura Leante, and Simone Tarsitani.
}
\keyword{datasets}
