% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{open_local_recordings}
\alias{open_local_recordings}
\title{Show downloaded recordings}
\usage{
open_local_recordings(folder_in = "Original", path = "~/movementsync")
}
\arguments{
\item{folder_in}{input folder relative to recording home (default is 'Original').}

\item{path}{recording home folder.}
}
\description{
Show downloaded recordings
}
\examples{
open_local_recordings()
}
