% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asynchrony.R
\name{pairs_title}
\alias{pairs_title}
\title{Helper to add title to pairs plot}
\usage{
pairs_title(recording)
}
\arguments{
\item{recording}{\code{Recording} object.}
}
\description{
Helper to add title to pairs plot
}
