% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot.Duration}
\alias{plot.Duration}
\title{Plot a Duration S3 object}
\usage{
\method{plot}{Duration}(x, ...)
}
\arguments{
\item{x}{S3 object}

\item{...}{passed to \code{\link[=barplot]{barplot()}}}
}
\value{
a plot object with durations.
}
\description{
Plot a Duration S3 object
}
\examples{
r <- get_sample_recording()
d <- get_duration_annotation_data(r)
plot(d)
}
