% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot.OnsetsSelected}
\alias{plot.OnsetsSelected}
\title{Plot a OnsetsSelected S3 object}
\usage{
\method{plot}{OnsetsSelected}(x, instrument = "Inst", tactus = "Matra", ...)
}
\arguments{
\item{x}{S3 object.}

\item{instrument}{column name.}

\item{tactus}{beat column name (defaults to "Matra").}

\item{...}{passed to \code{\link[=barplot]{barplot()}}.}
}
\value{
Return an 'OnsetsSelected' object.
}
\description{
Plot a OnsetsSelected S3 object
}
\examples{
r <- get_sample_recording()
o <- get_onsets_selected_data(r)
plot(o)
}
