% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot.View}
\alias{plot.View}
\title{Plot a View S3 object}
\usage{
\method{plot}{View}(x, columns = NULL, maxpts = 1000, ...)
}
\arguments{
\item{x}{S3 object}

\item{columns}{names of columns}

\item{maxpts}{maximum number of points to plot.}

\item{...}{passed to \code{\link[=plot.zoo]{plot.zoo()}}}
}
\value{
a plot object.
}
\description{
Plot a View S3 object
}
\examples{
r <- get_sample_recording()
v <- get_raw_view(r, "Central", "", "Sitar")
plot(v, columns = "LEar_x")
}
