% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{sample_offset_splice}
\alias{sample_offset_splice}
\title{Randomly create matching segments from a splicing table without overlaps}
\usage{
sample_offset_splice(splicing_dfr, v, num_splices, rejection_list = list())
}
\arguments{
\item{num_splices}{}
}
\value{
list of splicing data.frames
}
\description{
Works by adding a random offset to each start time in the splice. Uses rejection
sampling to avoid overlaps with the input segments and a additional segments
from a list of splices.
}
\examples{
r1 <- get_recording("NIR_ABh_Puriya", fps = 25)
d1 <- get_duration_annotation_data(r1)
rv1 <- get_raw_view(r1, "Central", "", "Sitar")
splicing_df <- splice_time(d1, tier ='INTERACTION', comments = 'Mutual look and smile')
x <- sample_offset_splice(splicing_df, rv1, num_splices = 100)
}
