% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{sapply_column_spliceview}
\alias{sapply_column_spliceview}
\title{Apply summary function to the columns in each segment of a SpliceView object}
\usage{
sapply_column_spliceview(sv, FUN, simplify = TRUE, USE.NAMES = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Apply summary function to each data point column in a SplicedView and return list of output data.
Simplify list to matrix.
}
