% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{splice_time.list}
\alias{splice_time.list}
\title{Generate spliced timeline using a list}
\usage{
\method{splice_time}{list}(x, ...)
}
\arguments{
\item{x}{named list.}

\item{...}{ignored.}
}
\value{
a \code{Splice} object.
}
\description{
Generate spliced timeline using a list
}
\examples{
l <- list(a = c(0, 10), b = c(10, 20), c = c(20, 30))
splice_time(l)
}
\seealso{
Other splicing functions: 
\code{\link{clip_splice}()},
\code{\link{get_spliced_view}()},
\code{\link{is_splice_overlapping}()},
\code{\link{merge_splice}()},
\code{\link{splice_time.Duration}()},
\code{\link{splice_time.Metre}()},
\code{\link{splice_time.OnsetsDifference}()},
\code{\link{splice_time.View}()},
\code{\link{splice_time}()},
\code{\link{split.SplicedView}()}
}
\concept{splicing functions}
