% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{splice_time.Metre}
\alias{splice_time.Metre}
\title{Generate spliced timeline using a Metre object}
\usage{
\method{splice_time}{Metre}(x, window_duration, rhythms = NULL, ...)
}
\arguments{
\item{x}{\code{Metre} object.}

\item{window_duration}{duration of window around metre.}

\item{rhythms}{vector of talas to subset on.}

\item{...}{ignored.}
}
\value{
a \code{Splice} object.
}
\description{
Generate spliced timeline using a Metre object
}
\examples{
r <- get_sample_recording()
m <- get_metre_data(r)
splicing_df <- splice_time(m, window_duration = 1)
head(splicing_df)
}
\seealso{
Other splicing functions: 
\code{\link{clip_splice}()},
\code{\link{get_spliced_view}()},
\code{\link{is_splice_overlapping}()},
\code{\link{merge_splice}()},
\code{\link{splice_time.Duration}()},
\code{\link{splice_time.OnsetsDifference}()},
\code{\link{splice_time.View}()},
\code{\link{splice_time.list}()},
\code{\link{splice_time}()},
\code{\link{split.SplicedView}()}
}
\concept{splicing functions}
