% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{splice_time}
\alias{splice_time}
\title{S3 generic function to splice a timeline}
\usage{
splice_time(x, ...)
}
\arguments{
\item{x}{S3 object.}

\item{...}{passed to relevant method.}
}
\value{
a \code{Splice} object.
}
\description{
S3 generic function to splice a timeline
}
\seealso{
Other splicing functions: 
\code{\link{clip_splice}()},
\code{\link{get_spliced_view}()},
\code{\link{is_splice_overlapping}()},
\code{\link{merge_splice}()},
\code{\link{splice_time.Duration}()},
\code{\link{splice_time.Metre}()},
\code{\link{splice_time.OnsetsDifference}()},
\code{\link{splice_time.View}()},
\code{\link{splice_time.list}()},
\code{\link{split.SplicedView}()}
}
\concept{splicing functions}
