% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.analyze.wavelet}
\alias{summary.analyze.wavelet}
\title{Summarise an analyze.wavelet object}
\usage{
\method{summary}{analyze.wavelet}(object, v, ...)
}
\arguments{
\item{object}{\code{analyze.wavelet} object.}

\item{v}{\code{View} object}

\item{...}{ignored.}
}
\value{
data.frame
}
\description{
Summarise an analyze.wavelet object
}
\examples{
r <- get_sample_recording()
rv <- get_raw_view(r, "Central", "", "Sitar")
pv <- get_processed_view(rv)
w <- analyze_wavelet(pv, "Nose_x")
summary(w, pv)
}
