% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.Duration}
\alias{summary.Duration}
\title{Summarise Duration object}
\usage{
\method{summary}{Duration}(object, ...)
}
\arguments{
\item{object}{\code{Duration} object.}

\item{...}{ignored.}
}
\value{
data.frame
}
\description{
Summarise Duration object
}
\examples{
r <- get_sample_recording()
d <- get_duration_annotation_data(r)
head(summary(d))
}
