% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.Metre}
\alias{summary.Metre}
\title{Summarise Metre object}
\usage{
\method{summary}{Metre}(object, tempo = FALSE, ...)
}
\arguments{
\item{object}{\code{Metre} object.}

\item{tempo}{summarise tempo? (Default is FALSE).}

\item{...}{ignored.}
}
\value{
list of summaries.
}
\description{
Summarises the cycle length or temp for each Metre.
}
\examples{
r <- get_sample_recording()
m <- get_metre_data(r)
summary(m)
}
