% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.OnsetsSelected}
\alias{summary.OnsetsSelected}
\title{Summarise OnsetsSelected object}
\usage{
\method{summary}{OnsetsSelected}(object, ...)
}
\arguments{
\item{object}{\code{OnsetsSelected} object.}

\item{...}{ignored.}
}
\value{
list of summaries.
}
\description{
Summarise OnsetsSelected object
}
\examples{
r <- get_sample_recording()
o <- get_onsets_selected_data(r)
summary(o)
}
