% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.Recording}
\alias{summary.Recording}
\title{Summarise Recording object}
\usage{
\method{summary}{Recording}(object, ...)
}
\arguments{
\item{object}{\code{Recording} object.}

\item{...}{ignored.}
}
\value{
list
}
\description{
Summarise Recording object
}
\examples{
r <- get_sample_recording()
summary(r)
}
