% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.View}
\alias{summary.View}
\title{Summarise a View object}
\usage{
\method{summary}{View}(object, ...)
}
\arguments{
\item{object}{\code{View} object.}

\item{...}{ignored.}
}
\value{
summary of data.frame.
}
\description{
Summarise a View object
}
\examples{
r <- get_sample_recording()
rv <- get_raw_view(r, "Central", "", "Sitar")
pv <- get_processed_view(rv)
fv <- apply_filter_sgolay(pv, c("Nose", "RWrist", "LWrist"), n=19, p=4)
summary(rv)
summary(pv)
summary(fv)
}
