% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{visualise_sample_splices}
\alias{visualise_sample_splices}
\title{Visualise random splices}
\usage{
visualise_sample_splices(
  splicing_list,
  jv,
  overlay = TRUE,
  avoid_splice_dfr = data.frame(),
  unstack = FALSE
)
}
\arguments{
\item{splicing_list}{a list of \code{Splice} objects.}

\item{jv}{\code{JoinedView} object.}

\item{overlay}{overlay the segments for a density plot?}
}
\value{
a \code{ggplot} object.
}
\description{
Visualise random splices
}
