% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{apply_column_spliceview}
\alias{apply_column_spliceview}
\title{Apply summary function to the columns in each segment of a SpliceView object}
\usage{
apply_column_spliceview(sv, FUN, simplify = FALSE, USE.NAMES = FALSE, ...)
}
\arguments{
\item{sv}{\code{SplicedView} object.}

\item{FUN}{function to apply.}

\item{simplify}{see \code{\link[=sapply]{sapply()}}.}

\item{USE.NAMES}{see \code{\link[=sapply]{sapply()}}.}

\item{...}{passed to FUN.}
}
\value{
see \code{\link[=sapply]{sapply()}}.
}
\description{
Apply summary function to each data point column in a SplicedView and return list of output data.
}
