% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{autoplot.SpectralDensityView}
\alias{autoplot.SpectralDensityView}
\title{Autoplot a SpectralDensityView S3 object}
\usage{
\method{autoplot}{SpectralDensityView}(object, period_range = c(0, 10), colour = "blue", ...)
}
\arguments{
\item{obj}{\code{SpectralDensityView} object.}
}
\value{
a \code{ggplot} object.
}
\description{
Autoplot a SpectralDensityView S3 object
}
\examples{
r <- get_recording("NIR_ABh_Puriya", fps = 25)
rv <- get_raw_view(r, "Central", "", "Sitar")
pv <- get_processed_view(rv)
sd1 <- spectral_density(pv, data_points = c("LElbow", "LEye"), spans = 5)
autoplot(sd1)

fv <- apply_filter_sgolay(pv, data_points = c("LElbow", "LEye"), n = 19, p = 4)
sd2 <- spectral_density(fv, data_points = c("LElbow", "LEye"), spans = c(3, 3))
autoplot(sd2)
}
