% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{difference_onsets}
\alias{difference_onsets}
\title{Get onset differences}
\usage{
difference_onsets(onset_obj, instruments, expr = NULL, splicing_dfr = NULL)
}
\arguments{
\item{splicing_dfr}{}
}
\description{
Get onset differences
}
\examples{
r1 <- get_recording("NIR_ABh_Puriya", fps = 25)
o1 <- get_onsets_selected_data(r1)
head(difference_onsets(o1, instruments = c('Inst', 'Tabla')))
head(difference_onsets(o1, instruments = c('Inst', 'Tabla'), expr = 'Matra == 3'))
}
