% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_metre_data}
\alias{get_metre_data}
\title{Get metre files}
\usage{
get_metre_data(recording, filetype = "rda", verbose = FALSE)
}
\arguments{
\item{recording}{\code{Recording} object.}

\item{filetype}{type of file ('rda' as default), can be 'csv'.}

\item{verbose}{messages the specific data loaded (default is 'FALSE').}
}
\value{
list of data.frames.
}
\description{
Get metre files
}
\examples{
r <- get_sample_recording()
m <- get_metre_data(r)
}
\seealso{
Other data functions: 
\code{\link{apply_filter_sgolay}()},
\code{\link{get_data_points}()},
\code{\link{get_duration_annotation_data}()},
\code{\link{get_feature_data}()},
\code{\link{get_filtered_views}()},
\code{\link{get_joined_view}()},
\code{\link{get_onsets_selected_data}()},
\code{\link{get_processed_views}()},
\code{\link{get_processed_view}()},
\code{\link{get_raw_optflow_view}()},
\code{\link{get_raw_views}()},
\code{\link{get_raw_view}()},
\code{\link{get_recording}()},
\code{\link{get_sample_recording}()}
}
\concept{data functions}
