% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{get_osf_recordings}
\alias{get_osf_recordings}
\title{Get movementsync recording from OSF}
\usage{
get_osf_recordings(
  stems = c("NIR_ABh_Puriya", "NIRP1_VS_Hams", "NIRP1_MAK_Jaun", "Gagaku_5_Juha",
    "NIR_DBh_Malhar"),
  to_dir = "~/movementsync/Original",
  overwrite = FALSE
)
}
\arguments{
\item{stems}{zip file stem(s) (default is stems for all the
recordings needed for the walk-throughs).}

\item{to_dir}{directory to copy to (default is "~/movementsync/Original").}

\item{overwrite}{overwriting existing dataset files?}
}
\value{
invisible vector of downloaded CSV file names.
}
\description{
Get movementsync recording from OSF
}
\examples{
\dontrun{
get_osf_recordings()
}
}
