% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{list_osf_recordings}
\alias{list_osf_recordings}
\title{List available recordings for movementsync from OSF}
\usage{
list_osf_recordings()
}
\value{
character vector of stem names
}
\description{
List available recordings for movementsync from OSF
}
\examples{
\donttest{
list_osf_recordings()
}
}
