% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{open_movementsync_data}
\alias{open_movementsync_data}
\title{Opens movementsync data home page at OSF}
\usage{
open_movementsync_data()
}
\value{
No return value, opens a browser on a specific OSF page
}
\description{
Opens movementsync data home page at OSF
}
