% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot.Metre}
\alias{plot.Metre}
\title{Plot a Metre S3 object}
\usage{
\method{plot}{Metre}(x, ...)
}
\arguments{
\item{x}{S3 object.}

\item{...}{ignored.}
}
\value{
a plot object with metre.
}
\description{
Plot a Metre S3 object
}
\examples{
r <- get_sample_recording()
m <- get_metre_data(r)
plot(m)
}
