% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{sapply_column_spliceview}
\alias{sapply_column_spliceview}
\title{Apply summary function to the columns in each segment of a SpliceView object}
\usage{
sapply_column_spliceview(sv, FUN, simplify = TRUE, USE.NAMES = TRUE, ...)
}
\arguments{
\item{sv}{\code{SplicedView} object.}

\item{FUN}{function to apply.}

\item{simplify}{see \code{\link[=sapply]{sapply()}}.}

\item{USE.NAMES}{see \code{\link[=sapply]{sapply()}}.}

\item{...}{passed to FUN.}
}
\value{
see \code{\link[=sapply]{sapply()}}.
}
\description{
Apply summary function to each data point column in a SplicedView and return list of output data.
Simplify list to matrix.
}
