% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{summary_onsets}
\alias{summary_onsets}
\title{Summary of difference in onsets}
\usage{
summary_onsets(
  onset_obj,
  recording,
  instruments,
  splicing_dfr = NULL,
  expr = NULL,
  show_plot = FALSE,
  filter_pair = NULL,
  na_omit = TRUE
)
}
\arguments{
\item{na_omit}{}
}
\description{
Summary of difference in onsets
}
\examples{
r1 <- get_recording("NIR_ABh_Puriya", fps = 25)
o1 <- get_onsets_selected_data(r1)
d1 <- get_duration_annotation_data(r1)
splice_dfr <- splice_time(d1, tier = 'Event', comments = 'tabla solo')
summary_onsets(o1, r1, instruments = c('Inst', 'Tabla'),
  splicing_dfr = splice_dfr, show_plot = TRUE)
}
